requirements.txt
matplotlib
django==1.11.2
gunicorn==19.6.0
djangorestframework
